Include python 3.4 and 3.5 to travis, include the acouchbase tests, make sure they do not fail on unsupported versions
